+++ src/main/php/net/stubbles/db/DatabaseConnection.php
@@ -32,6 +32,16 @@ interface DatabaseConnection
     public function details();
 
     /**
+     * returns property with given name or given default if property not set
+     *
+     * @param   string  $name
+     * @param   string  $default  optional  value to return if property not set
+     * @return  string
+     * @since   2.2.0
+     */
+    public function property($name, $default = null);
+
+    /**
      * establishes the connection
      *
      * @return  DatabaseConnection