+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+2.2.0 (2014-02-??)
+------------------
+
+  * introduce concept of properties for database config, allows arbitrary values
+     * added `net\stubbles\db\DatabaseConnection::property($name, $default)` to access connection config properties
+
 2.1.0 (2014-01-31)
 ------------------